backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
hvac==2.4.0
idna==3.11
psycopg2-binary==2.9.11
pycparser==3.0
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
